requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
anthropic>=0.40.0
